Update the UAuthority tests to catch non-binary IPs
The UAuthority proto file describes the IP field as holding "IPv4 or
IPv6 Address in byte format". The intent was for IPv4 addresses to be
represented in the native 32-bit binary representation (4 bytes) and
for IPv6 to be represented in the native 128-bit (16 byte) format.

Instead, BuildUAuthority is storing the human-readable string.

This change revises the tests for BuildUAuthority to catch the issue.