Add SafeMap and SafeUnorderedMap utility classes
SafeMap is a thread-safe wrapper around std::map, and SafeUnorderedMap
is a thread-safe wrapper around std::unordered_map.

Both implementations use a shared_mutex for protection against
concurrent modification. All const operations lock in the shared mode,
allowing for concurrent read access to the map, while anything that
would modify the map takes a unique lock to prevent read/write during
the modifying operation.

Interfaces returning or taking iterators are protected to guard against
excessive iterator invalidation and remove the risks involved in adding
locks to the iterators themselves (unintentionally holding an object
that itself is holding a lock).

To allow for bulk transactions and operations using iterators, the
transact() methods have been added. These methods take a std::function
(or lambda) that will be provided a reference to the base std::map or
std::unordered_map for use while the lock is held. These methods can
operate in either shared (const) or unique (non-const) modes.

For *many* uses of std::map and std::unordered_map, this should be a
drop-in replacement.