Reorganizing library headers
Moving things around to better align with the uProtocol spec's layering.

Layer 0: datamodel/ -> Interfaces for interacting with protobuf objects

Layer 1: transport/ -> The base classes for all transport protocol implementations

Layer 2: client/ -> The interfaces for uE clients on the uP network

Removes uri/ in favor of folding it in to the datamodel/ tree

Retains utils/ for common utilities that can be reused across transport
implementations.

Adding a readme file to the include/up-cpp directory providing guidance on the
folder structure.

Removing some unused headers