Spec update: uP spec is removing UPayload
This change converts all the client code to use the Payload builder
instead of UPayload. Because the payload format is now a message
attribute, it can be set on the UMessage builder. Once that happes, the
payload format will be enforced for all payloads used with that message
builder.

See up-spec#144 and up-spec#145