Implementing CallbackConnection utility
CallbackConnection.h provides a system where callback registrations can
be represented as a connection between two destructable handles. So long
as both handles are held, the connection remains valid and the caller
end can call the callback. If either handle is dropped, the connection
is severed and the caller will no longer be able to invoke the callback.

The implementation here aims to be as careful as possible around
concurrency, considering the possibility that callbacks may be invoked
and running in one context at the same time another context tries to
drop the callee's handle.

This utility is used anywhere callbacks are registered in up-cpp so that
the up-cpp API can be certain that discarded connections are always
cleaned up, regardless of whether the user of the API remembers to
explicitly drop the handle or not.