Rolling back to Conan 2.3.2
Conan 3.4.0 was released today, and it fixed some bugs that we were
using in our recipes. See https://github.com/conan-io/conan/pull/16272

Additionally, the old way of passing the conan cache was brittle, so it
has been switched to an artifact.

Lastly, a fixed conan profile is set to reduce the risk of changes
breaking things.