Remove unused fields from RpcServer
RpcServer::create() will create a lambda that wraps around the user's
callback. That wrapper will be registered with UTransport. Since the
wrapper is already a lambda, all the parameters it needs to produce
replies can be captured. As such, there is no need for the RpcServer
instance itself to hold those values.

This simplifies the code for implementing RpcServer and reduces the
risks of mistakes around setup.