Add unit test for UMessage validator isValid()
There was no check that the reasons returned by isValid() made sense or
were correct. It turns out that in most cases, except when evaluating a
notification type message, the only reason that was retured was
WRONG_MESSAGE_TYPE. This test now checks that the reasons are correct
for isValid().