Fixing expiration timing bug in RpcClient
Adjusts the output of the predicate used for evaluating if the
ExpireWorker should stop waiting on the condition variable such that it
will exit if the wake time of the request at the top of the priority
queue has changed since the worker started waiting. Also added check for
the queue becoming empty.