Revising UTransport API to fix spec alignment bug
The UTransport API was not well aligned with the final version of the
UProtocol 1.6.0 spec, in particular around the optionality of URI
filters for registered listeners. This change adds new methods that
better align with the intent of the spec. The existing APIs are retained
(in some cases with new workarounds for buggy behavior) having now been
flagged as deprecated.