Revising the UUri validator API per UTransport changes
UTransport now calls it the "entity" or "default entity" URI to avoid
confusion since it is not solely used as a source address (it could also
be used in constructing a sink filter).

Deprecating isValid() check. Either a specific message type validity
check should be used (in the case of a message with attributes) or the
isValidFilter() should be used.