RPC client/server test using UTransportMock
This test demonstrates up-cpp RPC code in a round trip test using
UTransportMock. This test is not working with a single transport
handle. It will work with two transport handles sharing the same
identity. It not clear if this is a problem with UTransportMock or an
unexpected issue in up-cpp code.