Adding exception type for empty callback functions (reopening of PR #301) (#304)
* Adding exception type for empty callback functions

This exception will be thrown when an empty std::function is passed to
the callback connection module. Adding the type first so that tests can
be written.

* Add callback tests for empty function objects

* Fix null callback assertions in L2 tests

Some L2 tests were asserting that the std::bad_function_call exception
should be thrown when a bad callback is called. We will be changing that
behavior so that the exception is thrown when the callback connection is
established, requiring updated test cases.

* Add empty function checks in CalleeHandle

The CalleeHandle class is the "owner" of the callback function objects
associated with a callback connection. As such, it will check the
validity of those function objects and throw the EmptyFunctionObject
exception if either the callback or cleanup function is empty / invalid.

* Update include/up-cpp/utils/CallbackConnection.h

Co-authored-by: Pete LeVasseur <plevasseur@gmail.com>

* specify Clang version 17 in CI

---------

Co-authored-by: Greg Medding <gregory.medding@gm.com>
Co-authored-by: Pete LeVasseur <plevasseur@gmail.com>